@@ -1,8 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Exports;
 
-use App\Client;
+use App\Models\Client;
 use Maatwebsite\Excel\Concerns\FromCollection;
 use Maatwebsite\Excel\Concerns\WithHeadings;
 
@@ -11,7 +13,6 @@ class ClientsExport implements FromCollection, WithHeadings
     public function collection()
     {
         return Client::select(
-
             'name',
             'email',
             'adresse',
@@ -25,14 +26,13 @@ class ClientsExport implements FromCollection, WithHeadings
     public function headings(): array
     {
         return [
-
             'Nom',
             'Email',
             'Adresse',
             'Matricule',
             'Téléphone 1',
             'Téléphone 2',
-            'Date de création'
+            'Date de création',
         ];
     }
 }
