@@ -1,14 +1,20 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Http\Controllers;
 
 use App\Exports\ClientsExport;
 use Maatwebsite\Excel\Facades\Excel;
+use Symfony\Component\HttpFoundation\BinaryFileResponse;
 
 class ClientExportController extends Controller
 {
-    public function export()
+    /**
+     * Export clients to Excel.
+     */
+    public function export(): BinaryFileResponse
     {
-        return Excel::download(new ClientsExport, 'clients.xlsx');
+        return Excel::download(new ClientsExport(), 'clients.xlsx');
     }
 }
