@@ -1,23 +1,31 @@
 <?php
 
-namespace App\Http\Controllers;
-use App\Imports\ClientsImport;
+declare(strict_types=1);
 
+namespace App\Http\Controllers;
+
+use App\Imports\ClientsImport;
 use Illuminate\Http\Request;
 use Maatwebsite\Excel\Facades\Excel;
+
 class ImportExportController extends Controller
 {
-    // export excel function
+    /**
+     * Import data from Excel file based on slug.
+     */
+    public function Import(Request $request, string $slug)
+    {
+        $request->validate([
+            'file' => 'required|file|mimes:xlsx,xls,csv',
+        ]);
 
+        if ($slug === 'clients') {
+            Excel::import(new ClientsImport(), $request->file('file'));
+        }
 
- // importation Excel
-    public function Import(Request $request , $slug){
-
-        if($slug == 'clients')
-            Excel::import(new ClientsImport, $request->file);
-
-
-        return back();
+        return back()->with([
+            'message' => 'Import terminé avec succès',
+            'alert-type' => 'success',
+        ]);
     }
-
 }
