@@ -1,44 +1,51 @@
 <?php
 
+declare(strict_types=1);
+
 namespace App\Models;
 
-use Illuminate\Contracts\Auth\MustVerifyEmail;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Foundation\Auth\User as Authenticatable;
 use Illuminate\Notifications\Notifiable;
 use Laravel\Sanctum\HasApiTokens;
 
-class User extends \TCG\Voyager\Models\User
+class User extends Authenticatable
 {
     use HasApiTokens, HasFactory, Notifiable;
 
-    /**
-     * The attributes that are mass assignable.
-     *
-     * @var array<int, string>
-     */
     protected $fillable = [
         'name',
-        'email','phone',
+        'email',
+        'phone',
         'password',
+        'role_id',
+        'avatar',
     ];
 
-    /**
-     * The attributes that should be hidden for serialization.
-     *
-     * @var array<int, string>
-     */
     protected $hidden = [
         'password',
         'remember_token',
     ];
 
+    protected function casts(): array
+    {
+        return [
+            'email_verified_at' => 'datetime',
+            'password' => 'hashed',
+        ];
+    }
+
     /**
-     * The attributes that should be cast.
-     *
-     * @var array<string, string>
+     * Check if user can access the admin panel.
+     * Will implement FilamentUser interface once Filament is installed.
      */
-    protected $casts = [
-        'email_verified_at' => 'datetime',
-    ];
+    public function canAccessPanel($panel): bool
+    {
+        return in_array($this->role_id, [1, 3]);
+    }
+
+    public function isAdmin(): bool
+    {
+        return $this->role_id === 1;
+    }
 }
