@@ -10,7 +10,7 @@
 
 <body>
     @php
-        $coordonnee = App\Coordinate::first();
+        $coordonnee = App\Models\Coordinate::first();
     @endphp
 
     <style>
@@ -178,7 +178,7 @@
             <div class="text-right">
                 <button id="printInvoice" class="btn btn-info" onclick="print()"><i class="fa fa-print"></i>
                     Imprimer</button>
-                <a class="btn btn-info" href="{{ route('voyager.tickets.index') }}"><i class="fa fa-file-pdf-o"></i>
+                <a class="btn btn-info" href="{{ route('admin.tickets.index') }}"><i class="fa fa-file-pdf-o"></i>
                     Retour</a>
             </div>
             <hr>
@@ -186,7 +186,7 @@
         <div class="container">
 
             <div class="receipt_header">
-                <img src="{{ Voyager::image($coordonnee->logo_facture) }}" data-holder-rendered="true"
+                <img src="{{ asset('storage/' . $coordonnee->logo_facture) }}" data-holder-rendered="true"
                     style="width : 220px ;     margin: auto;
                 display: block;
                 float: none;" />
